Rearranged types and adjusted namespaces
