+++ src/ConfigInjector/Configuration/ConfigurationConfigurator.cs
@@ -1,5 +1,5 @@
 ﻿using System.Reflection;
-using ConfigInjector.TypeProviders;
+using ConfigInjector.Infrastructure.TypeProviders;
 
 namespace ConfigInjector.Configuration
 {
