+++ src/ConfigInjector/Infrastructure/SettingValueConverter.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using ConfigInjector.Exceptions;
-using ConfigInjector.ValueParsers;
+using ConfigInjector.Extensions;
+using ConfigInjector.Infrastructure.ValueParsers;
 
-namespace ConfigInjector
+namespace ConfigInjector.Infrastructure
 {
     internal class SettingValueConverter
     {
@@ -22,13 +23,12 @@ namespace ConfigInjector
         private static IEnumerable<IValueParser> LoadValueParsers()
         {
             var parsers = Assembly.GetExecutingAssembly()
-                .GetExportedTypes()
-                .Where(t => typeof (IValueParser).IsAssignableFrom(t))
-                .Where(t => !t.IsInterface)
-                .Where(t => !t.IsAbstract)
-                .Select(t => (IValueParser) Activator.CreateInstance(t))
-                .OrderBy(vp => vp.SortOrder)
-                .ToArray();
+                                  .GetExportedTypes()
+                                  .Where(t => t.IsAssignableTo<IValueParser>())
+                                  .Where(t => t.IsInstantiable())
+                                  .Select(t => (IValueParser) Activator.CreateInstance(t))
+                                  .OrderBy(vp => vp.SortOrder)
+                                  .ToArray();
             return parsers;
         }
