+++ src/ConfigInjector/Infrastructure/TypeProviders/AssemblyScanningTypeProvider.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace ConfigInjector.TypeProviders
+namespace ConfigInjector.Infrastructure.TypeProviders
 {
     public class AssemblyScanningTypeProvider : ITypeProvider
     {
