+++ src/ConfigInjector/Infrastructure/TypeProviders/FilteredAssemblyScanningTypeProvider.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace ConfigInjector.TypeProviders
+namespace ConfigInjector.Infrastructure.TypeProviders
 {
     public class FilteredAssemblyScanningTypeProvider : ITypeProvider
     {
