+++ src/Sample.IntegrationTests/ConfigurationSettings/PersonNameValueParser.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using ConfigInjector;
 using ConfigInjector.Exceptions;
-using ConfigInjector.ValueParsers;
+using ConfigInjector.Infrastructure.ValueParsers;
 
 namespace Sample.IntegrationTests.ConfigurationSettings
 {
